various WASM terrain fixes

+++ src/addons/WorldGenerator/terrain/VRTerrain.h
@@ -74,7 +74,7 @@ class VRTerrain : public VRGeometry, public VRWorldModule {
         //void setupGeo();
         void setupMat();
         void createMesh(VRGeoData& geo, int res);
-        bool createMultiGrid(VRCameraPtr cam, int res);
+        bool createMultiGrid(VRCameraPtr cam, double res);
 
         void btPhysicalize();
         void vrPhysicalize();
